requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-documentai==3.6.0
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
idna==3.10
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.5.0
